refactor: 재귀적으로 playGame을 실행시키던 로직 삭제
재귀적으로 처리했을 때 값을 덮어쓰는 문제와 이벤트 리스너에 인라인 함수를 등록하는 부분에서 문제가 발생
