+++ src/game/playGame.js
@@ -1,14 +1,9 @@
-import printResult from './printResult.js';
-import getRacingCount from '../input/getRacingCount.js';
 import toggleRaceNumberInputScreen from '../init/toggleRaceNumberInputScreen.js';
+import getRacingCount from '../input/getRacingCount.js';
 
-export default function playGame(carList = '', value = 0) {
-  if (carList) {
-    toggleRaceNumberInputScreen();
-    getRacingCount(carList);
-  }
-  if (carList && value) {
-    toggleRaceNumberInputScreen();
-    printResult(carList, value);
-  }
+export default function playGame(carList) {
+  const $racingCountSubmit = document.querySelector('#racing-count-submit');
+
+  $racingCountSubmit.addEventListener('click', () => getRacingCount(carList));
+  toggleRaceNumberInputScreen();
 }
