+++ src/index.js
@@ -1,9 +1,7 @@
-import playGame from './game/playGame.js';
 import initializeGame from './init/initializeGame.js';
 
 export default function RacingCarGame() {
   initializeGame();
-  playGame();
 }
 
 new RacingCarGame();
