+++ src/init/initializeGame.js
@@ -1,7 +1,7 @@
 import addTagID from './addTagID.js';
 import addEventListener from './addEventListener.js';
-import toggleRaceNumberInputScreen from './toggleRaceNumberInputScreen.js';
 import toggleResultScreen from './toggleResultScreen.js';
+import toggleRaceNumberInputScreen from './toggleRaceNumberInputScreen.js';
 
 export default function initializeGame() {
   addTagID();
