+++ src/input/getRacingCount.js
@@ -1,4 +1,5 @@
 import playGame from '../game/playGame.js';
+import printResult from '../game/printResult.js';
 
 function checkValidRacingCount(racingCount) {
   if (Math.sign(racingCount) === 1) {
@@ -7,21 +8,12 @@ function checkValidRacingCount(racingCount) {
   return false;
 }
 
-function getRacingCountValue(carList, recursiveCount) {
+export default function getRacingCount(carList) {
   const $racingCountInput = document.querySelector('#racing-count-input');
   const racingCount = $racingCountInput.value;
 
   if (checkValidRacingCount(racingCount)) {
-    return playGame(carList, Number(racingCount), recursiveCount + 1);
+    return printResult(carList, Number(racingCount));
   }
   return alert('잘못 된 입력입니다! 0이상의 숫자만 입력 가능합니다');
-}
-
-export default function getRacingCount(carList, recursiveCount) {
-  if (!recursiveCount) {
-    const $racingCountSubmit = document.querySelector('#racing-count-submit');
-    $racingCountSubmit.addEventListener('click', () =>
-      getRacingCountValue(carList, recursiveCount),
-    );
-  }
 }
